Simply ignore exceptions in child threads now (those are regularly restarted anyway).
Apparently, the asynchronicity actually works as it should now, though the whole program doesn't feel as smooth as we'd like in case of hard-to-calculate functions. Also, when allowing for actual threads / multi-core, the interface starts acting funny, text is again box-replaced, even segfaults have been occured. This could certainly be fixed. But staying responsive is the more important thing right now.