Hack to get rid of strange deadlock in some processor-intensive plots.
When using `plotLatest` on results of an expensive iterative computation in single-threaded runtime mode,
the entire program would sometimes freeze. Rather ominously, this problem disappears when mandating
a small minimum thread wait-time (1 ms in this case).

It should at some point be investigated what's causing this problem in the first place.